countries where driving automobiles on the right is the rule (that is, USA, China and most other countries except for Great Britain, Australia, and some former British colonies), multi-lane freeways often employ**\*a rule that requires drivers to drive in the right-most lane unless they are passing another vehicle, in which case they move one lane to the left, pass, and return to their former travel lane.*** Build and analyze a mathematical model to analyze the performance of this rule in light and heavy traffic. You may wish to examine tradeoffs between traffic flow and safety, the role of under- or over-posted speed limits (that is, speed limits that are too low or too high), and/or other factors that may not be explicitly called out in this problem statement. Is this rule effective in promoting better traffic flow? If not, suggest and analyze alternatives (to include possibly no rule of this kind at all) that might promote greater traffic flow, safety, and/or other factors that you deem important.
In countries where driving automobiles on the left is the norm, argue whether or not your solution can be carried over with a simple change of orientation, or would additional requirements be needed.

Lastly, the rule as stated above relies upon human judgment for compliance. If vehicle transportation on the same roadway was fully under the control of an intelligent system ¿ either part of the road network or imbedded in the design of all vehicles using the roadway ¿ to what extent would this change the results of your earlier analysis?